backend/requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
flask-cors==6.0.1